Desktop/shadz-os/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
psutil==6.1.1
pydantic==2.11.3
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.0.1
